Remove event_msg dependencies from event_parsing mod
